tunneling: Access correct IP header when processing ECN.
We attempt to copy the ECN bits from the outside of the tunnel to
the inside on receive if we are encapsulating IP traffic.  However,
we were previously looking at the inner IP header as the source of
the ECN bits, when it should have been the outer header.  This
corrects that and cleans up the function a little bit.

Signed-off-by: Jesse Gross <jesse@nicira.com>
Acked-by: Ben Pfaff <blp@nicira.com>