datapath: Don't use RCU for internal dev vport.
The vports are now attached and ready to go when they are allocated,
so we don't have to worry about future changes.  As a result, we can
directly store the pointer in the internal dev's netdevice private
space before it is registered.  The registration process will handle
the necessary write memory barriers and anyone who has a reference
to the netdev will have done the read side barriers, we don't need
to use RCU at all.

Signed-off-by: Jesse Gross <jesse@nicira.com>
Acked-by: Ben Pfaff <blp@nicira.com>