Makefile: Check for undistributed files on every make, not just "make dist".
It's really easy to add files to the Git repository but forget to add them
to the distributions created by "make dist".  I do this regularly, for
example.  For some time, we've had a check that runs on "make dist" to
make sure that the distribution is complete, but I still screw up because
I don't run "make dist" all that often.

This commit improves the situation, by doing the check on every "make",
instead of just on "make dist".